backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google_auth_oauthlib==1.2.1
google-api-python-client
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
Flask-CORS==4.0.1
